Top-level-await with TypeScript
this is an experimental config to allow TypeScript and Node understand ECMAScript modules using ts-node, implementing "type": module, in the package.json.

Thanks to this configuration I didn't have to do a refactor of the code base, adding the extension file to all the imports.

I can do dynamic imports with top-level-await, but now is not possible to use require. It's the price to achieve the benefits of the top-level-await.